+++ src/vmm/src/devices/virtio/block/virtio/test_utils.rs
@@ -43,6 +43,7 @@ pub fn default_block_with_path(path: String, file_engine_type: FileEngineType) -
         is_root_device: false,
         partuuid: None,
         is_read_only: false,
+        discard: false,
         cache_type: CacheType::Unsafe,
         // Rate limiting is enabled but with a high operation rate (10 million ops/s).
         rate_limiter: Some(RateLimiterConfig {