+++ tests/framework/microvm.py
@@ -910,6 +910,7 @@ class Microvm:
         partuuid=None,
         cache_type=None,
         io_engine=None,
+        discard=None,
     ):
         """Add a block device."""
 
@@ -922,6 +923,7 @@ class Microvm:
             partuuid=partuuid,
             cache_type=cache_type,
             io_engine=io_engine,
+            discard=discard,
         )
         self.disks[drive_id] = path_on_host